+++ src/components/BlankView.jsx
@@ -6,15 +6,14 @@ import Header from '@stevederico/skateboard-ui/Header';
  * Minimal view with Header only. Use as starting point for new views.
  *
  * @component
+ * @param {Object} props
+ * @param {string} props.title - Header title
  * @returns {JSX.Element} Blank view with header
  */
-export default function BlankView() {
+export default function BlankView({ title = "Blank" }) {
   return (
     <>
-      <Header
-        buttonClass=""
-        title={"Blank"}
-      />
+      <Header title={title} />
     </>
   )
 }